Add "missing" semicolon to parser's generator.mly
Summary:
This doesn't appear to be a bug, parenthesized yeild statements parse
just fine in the current code - I'm thinking maybe the .mly syntax
is forgiving and will accept missing `;` when the grammar is unambiguous.

The test case proving that this works prior to my change is here:
https://github.com/facebook/pyre-check/blob/main/source/parser/test/generatorTest.ml#L4380-L4389

And I found this blurb in menhir release notes suggesting that semicolons
are indeed optional and ignored:
https://gitlri.lri.fr/ftuong/menhir/blob/51abc4bbbb4cea5d09dce11dd2805b7912681542/CHANGES.md#anchor-20181026

But it's very confusing, we use `;` to separate the tokens in every other
form that I can see; it looked like a bug at first to me (I'm investigating
an actual yield-related parsing bug whose root cause turns out to be
more complicated). We should try to be consistent.

Reviewed By: grievejia

Differential Revision: D30815477

fbshipit-source-id: 9d909ce3f6ab173d35631539ae18cb2ea7f11d10