PEP 342 [syntax][3/n] forward_expression
Summary:
**Stack description**

The TL;DR of this stack is that we need to get rid of yield statements entirely,
and work with yield only as expressions, in order to
 (a) correctly support PEP 342
 (b) unblock switching to the CPython parser and pyre-ast

For context on the stack as a whole, and how the various stages
fit together, check D30840005 / the commit labeled "PEP 342 [syntax][0/n]"

**Commit description**

This commit does part of the typeCheck.ml work: I'm hoisting our current attempt
to use an `__iter__` call to resolve the meaning of a `yield from` out of parsing and
into typechecking, where it belongs.

Specifically, I'm reproducing this logic here, which is (in the parser!!) doing
a magic rewrite of `yield from` expressions:
https://github.com/facebook/pyre-check/blob/1a89f78/source/parser/generator.mly#L609-L623

We don't want to do this rewrite going forward - it's wrong
in that it drops necessary information, and it diverges from the official
python Ast - which means we need that logic to move into typecheck.

The next commit will change `forward_statement`; I probably could have
squashed them into a single commit but this was my first major change to
`typeCheck.ml` and I wanted to keep commits as small as possible.

Reviewed By: grievejia

Differential Revision: D30840507

fbshipit-source-id: 778e6b4de362e9f5519c7c0041a3c5b02d1c758e