PEP 342 [syntax][5/n] Cut over the parser
Summary:
**Stack description**

The TL;DR of this stack is that we need to get rid of yield statements entirely,
and work with yield only as expressions, in order to
 (a) correctly support PEP 342
 (b) unblock switching to the CPython parser and pyre-ast

For context on the stack as a whole, and how the various stages
fit together, check D30840005 / the commit labeled "PEP 342 [syntax][0/n]"

**Commit description**

This commit is where we actually change production: the previous
commits added a YieldFrom expression and supported it in
`typeCheck.ml` in a way that doesn't require an Ast rewrite,
but up until this diff we would never actually parse code
into the new Ast.

We update the preprocessing logic at the same time as the parsing
logic because they go hand-in-hand; the major change here is that
we should look for yield form *expressions* rather than *statements*
when deciding whether to add implicit returns (the previous logic
was actually incorrect anyway, it would miss non-toplevel yield
expressions entirely).

I'm still leaving the Statement.{Yield,YieldFrom} variants in
this commit, but without any typechecking support, because the
commit where I remove them is going to involve a lot of easy
boilerplate changes to match statements and I don't want
to clutter the critical parser change with that.

**Some detailed notes - parsing PEP 342 is tricky**

Parsing yield expressions in a way that agrees with PEP 342
is pretty annoying because the parentheses rules mean that
even though yeild is always syntactically an expression, we
cannot directly write the parse rules as an expression case:
- yield forms surrounded by parentheses are always valid
  expressions
- but yield forms without parentheses are only valid expressions
  in two cases
  - when used as bare statements for just side effects
  - when used in the right-hand-side of an assignment

To support this, we:
- have a grammar rule `yield form` representing either a `yield` or `yield
  from` expression (we already did this with the old `yield` grammar
  rule, but it was incorrect because we did not have YieldFrom expressions)
- always take a `yield_form` surrounded by parentheses in the
  `expression` rule
- always take a `yield_form` without parentheses in the `value` rule
- make sure that `short_statement` fully supports using `value`
  - in bare expressions (it already did)
  - in assignments with no annotation or comment annotation
    - it did *not* because the existing rule that we hit involved
      a `test_list`, which doesn't match non-expression `value` rules
    - so, I added a new case for bare `values` which gets matched
      after the existing `test_list` case
  - in assignments involving bare `EQUALS` with annotations
    - it did *not*, again because of matching `test_list`
    - so again I added a new case after the existing `test_list` one
      to match `value`.

Reviewed By: grievejia

Differential Revision: D30841859

fbshipit-source-id: b69ab955b924da3dde02204f1578b751b9be7ab2