PEP 342 [syntax][0/n] rearrange code
Summary:
This commit moves the `Statement.Yield*`  logic so that it
lives right next to the `Statement.Expression` in typeCheck.ml.

The reason to do so is that in the rest of the stack, we are going to replace
yield statements with expressions entirely, and this will require migrating the
logic out of `Statement.Yield*` cases into `Statement.Expression` cases that
match on the inner expression having a yield form.

It's easier to both write and review the changes if we make the existing
code live next to one another.

# Context on the stack: pyre handles generator syntax incorrectly

In [PEP 342](https://www.python.org/dev/peps/pep-0342/#new-syntax-yield-expressions),
the `yield` syntax was redefined to be an expression with side effects rather
than a statement. We know that the CPython Ast in fact treats yields this
way - see for example the
[pyre-ast concrete AST](https://github.com/grievejia/pyre-ast/blob/50251b3/lib/concrete.ml#L276-L277)
for easily skimmable proof of this.

So, I am working on a change to fix our Ast. Once I'm done with a minimal
Ast fix I'll also try to address typechecking issues - obviously our typechecking
sometimes fails right now since we can't even express cases like returning
a `yield from` in our current incorrect AST.

The Ast fix is more urgent because it causes parse errors, which cannot be
addressed with fixme / ignore directives. It also will definitely block
our planned transition to using pyre-ast and the underlying CPython parser.

**To do _after_ this stack: fix typechecking errors**

Fixing typechecking will require some restructuring of `typeCheck.ml` because
we don't currently have a way to validate return types from expressions, which
is incorrect according to PEP 342. The gaps in typechecking are related to
but not the same as the Ast issues - for example we can parse this code just
fine but we fail to detect the type error because we completely ignore
yeilds that aren't in a top-level statement
```python
from typing import *

def f() -> Generator[int, None, None]:
    print((yield "hello"))  # typechecks, but it shouldn't!
```

We also incorrectly throw typechecking errors on generators that return
values, which fails to match parts of both PEP 342 and 380. The following
code, for example, produces an incorrect IncompatibleReturnType
```python
from typing import *

def f() -> Generator[int, None, str]:
   yield 1
   return "x"  # this is correct, the generator is marked as stopping
               # with a string value. But pyre complains.
```

We can fix the parsing failures by making both `Yield` and `YieldFrom` into
expressions and treating bare statements with `Yield*` expressions exactly
as we currently treat `Yield*` statements, and then handle the harder
typechecking problems in a separate stack.

**Outline of the upcoming commits**

I'm actually not sure whether people usually land expansive changes like
Ast modifications in one go to preserve a clean state or in easier-to-review
bits that result in intermediate states.

I'm opting for the latter because I usually find it easier for both
code review and history to keep boilerplate and tricky logic changes
as separate as possible. In the rest of this stack I will:
- first add the new expression to the Ast
  - this is easy boilerplate, it touches almost a dozen files with small changes
  - it can't break anything because the parser never produces a YieldFrom
  - it needs to be done first because the later logical changes require having
    something to match on
- update various easy static analyses (awaitable check, taint check) so that
   they treat Statement.Expression and Statement.Yield* the same way; that
   will make later changes transparent to them
- add the ability to match yield form expressions inside of `Statement.Expressions`
  the same way we currently match `Statement.Yield*` variants in typeCheck.ml.
  I'll do it in two small commits:
  - first update the `forward_expression` logic
  - then update the `forward_statement` logic
- flip the parser and preprocessor
  - this is when all the changes actually take affect
  - I also convert the typeCheck logic for `Statement.Yield*` into do-nothing
    stubs in this commit, because it guarantees that the old logic isn't
    somehow hiding bugs
  - note that the correctness of the logic changes in earlier steps isn't
    really testable until this commit - once we see tests pass here we know
    that we don't have any regressions
- clean up all references to `Statement.Yield*`
  - I patch up an (apparently untested) `pyre infer` case at this point
  - most of the rest of it is boilerplate
  - it turns out our CFG code is doing unnecessary things, which I learned
     when I did this step. I'll stack a separate commit (not directly tied to
     PEP 342) to clean that up.

Reviewed By: grievejia

Differential Revision: D30840005

fbshipit-source-id: 3e98357ea246984574ed759b108ae16eededab39