PEP 342 [syntax][2/n] Update awaitable/taint analyses
Summary:
**Stack description**

The TL;DR of this stack is that we need to get rid of yield statements entirely,
and work with yield only as expressions, in order to
 (a) correctly support PEP 342
 (b) unblock switching to the CPython parser and pyre-ast

For context on the stack as a whole, and how the various stages
fit together, check D30840005 / the commit labeled "PEP 342 [syntax][0/n]"

**Commit description**

In this commit, I update all of the non-typecheck static analysis
code that cares about yield handling to be ready for a world
of yield-forms-as-expressions-only.

The two analyses that are impacted are awaitable checks and
taint analysis. I'm fuzzy on why exactly awaitable check is
affected but it probably has to do with old-style async coroutines, and
it was pretty easy to update it just by looking at the existing logic.

Taint is more obviously affected because yields cause data to
jump between caller and callee in pretty much the same way as
returns and method calls do.

To make the update, in all cases we make the following changes:
- do all the yeild-specific logic out of yield statement
  cases and into yield expression cases
- move the yield statement cases up to live right next to the
  expression-statement case, and use exactly the same case-
  handling block of code. This is perfect since as soon as we
  flip the parser, these will *become* expression-statements.

Reviewed By: grievejia

Differential Revision: D30852585

fbshipit-source-id: c8ba46ba8350b65fede7386220d662507c24c5cf