Add `is_object_target` boolean flag to parse functions
Summary:
This allows us to have more descriptive error messages for when a standalone `ViaTypeOf` is used in a non-attribute model.

Before:
```
`ViaTypeOf` is an invalid taint annotation: No such parameter `$global`
```

After:
```
`ViaTypeOf` is an invalid taint annotation: A standalone `ViaTypeOf` without arguments can only be used in attribute or global models.
```

Reviewed By: arthaud

Differential Revision: D30841692

fbshipit-source-id: ddecf08ae645eee21d0b9b78987112672ebb0f76