Support `ViaTypeOf` on attributes
Summary:
We want to support this kind of syntax for `ViaTypeOf` on attributes:
```
C.x: ViaTypeOf = ...
```
which is equivalent to
```
C.x: TaintInTaintOut[ViaTypeOf] = ...
```

We also want to support:
```
C.x: TaintSource[Test, ViaTypeOf] = ...
C.x: TaintSink[Test, ViaTypeOf] = ...
```

**Note**: Tags are not currently supported, but maybe in the future.

Since attributes are modelled as fake 'targets' with a symbolic parameter `$global`, this change supports `ViaTypeOf[$global]` on attributes and makes the above syntaxes essentially syntactic sugar for this form. i.e. an annotation of `ViaTypeOf` is treated as `TaintInTaintOut[ViaTypeOf[$global]` and a standalone identifier
`ViaTypeOf` (not the call) in the list of kinds is treated as`ViaTypeOf[$global]` during model parsing.

Doing this requires minimal change to the existing code and putting the logic in kind parsing we don't have to special case the rewriting for `TaintSource`, `TaintSource`, and `TaintInTaintOut`. Another option is to have a preprocessing step during the parsing before everything gets sent to `parse_annotation`, but then we'd have to extract the `ViaTypeOf` from the list of kinds before passing the list along, create the breadcrumb, etc.

Reviewed By: arthaud

Differential Revision: D30770136

fbshipit-source-id: e7c7dabb655c19d04639ce8e7f39c0bf4070801f