CFG construction: handle while True/False
Summary:
Change the CFG for while, so that:
* in case of `while True`: do not connect `orelse` to join node.
* in case of `while False`: have a trivial body, and do not connect the body to the split node.

`while True` is really the case we care about; the `while False` is somewhat of a nice side-effect, and good to do for consistency.

This allows us to handle code like:
```
while True:
  ...
  if some_condition_becomes_true:
    break       # or return
  ...
```
properly.

In the scenario that the while loop body never breaks, we maintain behavior as before (i.e. allow exit of the loop). This keeps some of the assumptions the downstream code is written under happy, while still improving the handling of cases like above.

Reviewed By: grievejia

Differential Revision: D30884573

fbshipit-source-id: abf249e2ad3d9128b2b0be6a2a2ab5bcf42f7038