CFG construction: `assert False` jumps to error
Summary:
# Context
I think we should consider treating `True` and `False` specially in the CFG construction, when used as a "test expression" (`assert`, `while`, `if`).

These are common patterns found in the wild:

```
while True:
  ...
```
```
assert False, "msg"
```

The latter would generally appear under a conditional, e.g.:
```
if ...:
  ...
elif ...:
  ...
else:
  assert False, "msg"
```

As an example, in this commit, I show what I think the behavior could be for `assert False`. `assert False, msg` is equivalent to `raise AssertionError(msg)`, and in that should take one to the error state.

In the case of `while True` for instance; during AST construction:
```
    | { Ast.Node.value = While ({ While.test; body; orelse } as loop); _ } :: statements ->
        (*       _____________
         *       v           |
         *  -> [split] -> [body]
         *       |          :
         *       v          :
         *    [orelse] -> [join] -> *)
````
there is no need to have an edge from the `split` to `orelse`. This would remove issues such as no handling of assert False in uninitialized local check*, and also less special handling in typecheck.

In type check, pysa, uninitialized local check, and any other checks that depend on control flow, this will lead to:

1. consistent behavior
2. less duplicated effort

# This commit

This commit adds special handling, narrowly, for "assert False".

# Commits to follow in the stack

* Add handling for `if` and `while`. I should probably note that, from the point of view of CFG construction, `assert` handling is *not* the same as `if` or `while` (which we translate to an assert inside the blocks) . For example, an `assert False` leads to an error state, an `if False` just skips the block.

* Look into implementing reduction of constant boolean expressions (for example, have `x and False` behave similarly to `False`).

Reviewed By: grievejia

Differential Revision: D30740701

fbshipit-source-id: 04217c29b59a09b0fe71d91f7f381b32d680b9c1