AnnotationCollector: don't include self/cls parameters
Summary:
Up until D30799267 (https://github.com/facebook/pyre-check/commit/8653ef77fcd534a203f96de2b516a6d3237bc366), we were counting self as an annotated parameter.

For the coverage collector; we use the parameter annotations directly to mark a line as covered or not; so that didn't fix `coverage` command. This most egregious and commonly occurring case of that:

```
            class A:
                foo(42)
                def baz(self): ...   # Covered
```
Above was marked as covered earlier; and this diff fixes that.

It does leave one edge case still not handled correctly for coverage collection, namely a method where the self is annotated but the return value is not (why would someone do that?). In any case, that will show up as "Not covered" even though Pyre will type check it. Though that pattern is not common and might not warrant a fix (I won't be surprised if there are near-0 occurrences of that in our codebase), I think fix involves tracking implicitly annotated from annotated (which might be worth doing in its own right). In any case, I leave that for another day.

Reviewed By: shannonzhu

Differential Revision: D31033377

fbshipit-source-id: 0d8c35c7f43c23b2679b4e68b04b469a227ef922