Support ViaTypeOf in the DSL
Summary:
This adds support for `ViaTypeOf` in the DSL, mostly to be used when modeling parameters. Since we don't always know the name of the parameter that will be matched, an empty `ViaTypeOf` in the `Parameters` clause means that it'll be treated as a `ViaTypeOf` on the parameter it's on.

This is done by adding an additional flag allowing a standalone `ViaTypeOf` when parsing the model query. The model parser just parses it as a `ViaTypeOf` on the `$global` symbolic parameter, and then later when the production gets turned into an actual taint model, it rewrites the annotation to use the actual parameter instead of `$global`.

Reviewed By: arthaud

Differential Revision: D31120359

fbshipit-source-id: ff7735ed60b1a094b22938906754c33bbffc6a0c