Unit test creation of qualifier with extension suffix
Summary:
Configerator is running into a lot of false positives due to imports that include the suffix, ie., to import from `a.b.cinc` the statement looks like `from a.b.cinc import Foo`, which makes Pyre angry.

Trying to debug what caused the regression. Looks like the qualification of the module is still happening as expected, so just committing this unit test for now.

Reviewed By: dkgi

Differential Revision: D31162507

fbshipit-source-id: dfa35b799956ae6b8e5eeb121d691d11220082ad