Flush shared memory info from command line by logging module tracker build
Summary:
Super trivial change.

When experimenting with type checking in configerator, I was seeing servers seemingly hanging on:

```
2021-09-28 16:36:20,429 [PID 3062075] INFO Initializing shared memory [heap_size=8589934592, dep_table_pow=27, hash_table_pow=26]
```

which made it seem like something was wrong with shared memory or the heap.

In reality, building the module tracker can take a long time if the search path given is too large. In configerator in particular, we're actually blocked on simply passing in `configerator/source` as a search path because the module tracker would take tens of minutes to go through all the files.

 ---

Also potentially of interest (but not related to this diff), parsing the same number of files is 5x slower in configerator than in distillery. Building the environment was 3x slower. I wonder what the cause is.

```
2021-09-28 16:36:24,056 [PID 3062075] PERFORMANCE Module tracker built: 3.57s
2021-09-28 16:36:24,056 [PID 3062075] INFO Building type environment...
2021-09-28 16:36:24,312 [PID 3062075] INFO Parsing 37016 stubs and sources...
2021-09-28 16:37:18,378 [PID 3062075] PERFORMANCE Sources parsed: 54.01s
2021-09-28 16:38:20,951 [PID 3062075] PERFORMANCE Full environment built: 116.76s
2021-09-28 16:38:33,853 [PID 3062075] INFO Checking 528 functions...
2021-09-28 16:38:39,002 [PID 3062075] PERFORMANCE Check_TypeCheck: 5.14s
2021-09-28 16:38:39,003 [PID 3062075] MEMORY Shared memory size post-typecheck (size: 1946)
2021-09-28 16:38:58,903 [PID 3062075] PERFORMANCE Check: 158.24s
```
```
2021-09-28 17:59:34,469 [PID 288118] PERFORMANCE Module tracker built: 5.28s
2021-09-28 17:59:34,469 [PID 288118] INFO Building type environment...
2021-09-28 17:59:34,699 [PID 288118] INFO Parsing 69576 stubs and sources...
2021-09-28 17:59:52,601 [PID 288118] PERFORMANCE Sources parsed: 17.90s
2021-09-28 18:00:19,062 [PID 288118] PERFORMANCE Full environment built: 44.59s
2021-09-28 18:00:30,932 [PID 288118] INFO Checking 770565 functions...
2021-09-28 18:07:02,304 [PID 331177] PERFORMANCE Check_TypeCheck: 134.05s
2021-09-28 18:07:02,304 [PID 331177] MEMORY Shared memory size post-typecheck (size: 3171)
```

Reviewed By: dkgi

Differential Revision: D31253855

fbshipit-source-id: bfd2020e881a242fea915453279e15b8f2649e83