Opt for user independent server socket location in client
Summary:
Pass in optional `socket_path` field to server config, using the user independent hashing.

NOTE: will not land this diff until the binary changes have been landed *and* fully rolled out, to avoid the client looking for the socket in the user-independent location, and the binary falling back to the user-dependent default location.

Next steps:
1. land and roll out binary, then land this diff (current failing integration tests are expected)
2. rename/deprecate old socket path logic, making socket_path field mandatory

Reviewed By: grievejia

Differential Revision: D31104967

fbshipit-source-id: b8ecfeca7b88fe8418cb880718cb4d0fea5c95e4