Add a performance profiler in the taint analysis
Summary:
This patch adds a performance profiler specific to the taint analysis.
It can be enabled by adding a `pyre_dump_perf` statement in the function or method to profile.
This will dump extensive results about the analysis time spent into a callable, which will enable us to find performance bottlenecks.

Reviewed By: dkgi

Differential Revision: D31323057

fbshipit-source-id: c2a9129547f6fa60566e88532f4b30ac242623a3