Add check for untested `type(_) is _` refinement
Summary:
I began looking into what it would take to implement better type
refinement to support pytorch, and noticed that we already support
`type(x) is int` but it was easy to miss this because there was no
test case.

I've refactored our type comparison assertion tests (where I use the
term "comparison" to cover both `isinstance(_)` and `type(_)` based
comparisons) to use a helper function that verifies both
functionalities. With this testing change in place, we're ready to
add support for type *equality* checks, which at runtime behave
exactly like type `is` checks as far as I can tell but pyre does not
support yet.

Reviewed By: grievejia

Differential Revision: D31452916

fbshipit-source-id: 16ffa1c1b4dfa0eefff61e927d2138f2453208bc