Fix uninitialized local error.
Summary:
I'm actually not sure why our own uninitialized local check doesn't complain
here, but I discovered when spinning up some integration tests (which I had
to abandon) that the `persistent` command's `finally:` block is incorrectly
assuming any errors in the `try/except` happen after `request` is initialized.

This is false, the line defining `request` can fail which means we should
initialize `request` to None and log with `id` of `None` if `request` is
not defined.

If we omit this, we might never see the true exception which would be bad for
debugging.

Reviewed By: arthaud

Differential Revision: D31559121

fbshipit-source-id: 0f7b09c79179be38a1d5694614b1d2ecbc87920f