Reorder some refinement cases
Summary:
I want to add support for `type(_) == _` and `type(_) !=` to our
type comparison refinements.

At the moment the `callable` refinement has gotten mixed in with
the type comparison refinement... in both we have to have cases
for both positive and negated checks, but the actual refinement
logic is different.

To me it makes more sense to group the "same kind of" refinement
operations together, that will make it easier to see where to make
related positive and negative changes.

Reviewed By: arthaud

Differential Revision: D31559821

fbshipit-source-id: 8eea5302053966e01b9a0fd092946f94634aa1f6