Implement a weak topological order over the control flow graph
Summary:
# Problem
Our current fixpoint engine is quite inefficient, as it uses a simple breadth first search on the control flow graph.
This can lead to nodes being re-analyzed eagerly, especially in the backward analysis.

For instance:
```
y = "hello" # This is analyzed 4 times by the backward analysis.
if z == 1:
  return y
elif z == 2:
  return z
elif z == 3:
  return z + x
elif z == 4:
  return z + x
else:
  return x
```

# Solution
We should analyze nodes in the order provided by a weak topological order, as described in
F. Bourdoncle. Efficient chaotic iteration strategies with widenings.
In Formal Methods in Programming and Their Applications, pp 128-141.

This can also be used to find heads of loops, which is a very good candidate to apply widening and check for convergence.
Note that our current fixpoint algorithm applies widening and checks for convergence at the beginning of every nodes, which is also quite inefficient.

# Implementation
The implementation is unashamedly inspired by the SPARTA implementation, which is itself taken from the paper.
https://github.com/facebook/SPARTA/blob/main/include/WeakTopologicalOrdering.h#L267

# Follow-up
In a follow-up, we will use the weak topological order to implement our fixpoint engine.

Reviewed By: dkgi

Differential Revision: D31525099

fbshipit-source-id: 8a5a17466005f091d9efe134cbce6fa3e18edd03