Support `type(_) == _` style type refinement
Summary:
Adds support for `type(_) == _` and `type(_) !=` to our
type comparison refinements, as well as unit test coverage.

Because `type(x) == Y` and `type(x) is Y` have the same behavior
at runtime as far as I know, the cases can be handled exactly
the same.

Technically the behavior actually differs from `isinstance`, but
because at this point we only do forward analysis rather than constraint
solving the distinction isn't relevant to our existing algorithm.

In a constraint-solving world `isinstance` would place a bound on
the resolved type whereas `type(_) {is,==} _` would be an exact
match, and the negated behavior would be even more different. Worth
keeping in mind for later.

Reviewed By: kbansal

Differential Revision: D31560550

fbshipit-source-id: 6d00c344e81627331250c96eeb16d8a65cb4963e