Adds more details to ModelVerification error
Adds more details to model verification error by adding additional
pieces of code that wrap over the errors passed to the python from the
Ocaml code and see whether the module in the error is installed or not
using pkg_resources from setuptools.

This implementation is kinda hacky and is dependentant on the error
message (description) format of the ModelVerification errors. A suitable
way to tackle this would be to pass the module name as well from Ocaml.

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>