Remove location information from function names
Summary:
The CPython AST does not store location information
for the name portion of a function, so we need to remove this
in order to make migration to pyre-ast easy.

This commit is huge but pretty straightforward, I'm just unwrapping
the name from a `Reference.t Node.t` to a bare `Reference.t`

The visible API behavior should be 100% unchanged:
- the only places where we used location information were
  `visit.ml` and `location.ml`
- I wrote a helper function that computes the location of the name
  (which isn't stored in the CPython ast but is recoverable from the
  body location + the function name) so that those APIs don't change

we could later simplify by not insisting on computing the location of
the name but I think it's better if this change doesn't
affect behavior at all, since it's a pretty huge change in terms of raw
line count.

Reviewed By: dkgi

Differential Revision: D31579648

fbshipit-source-id: 55e72d67d0b01dec1bf968502f5a439bf1fc3c3e