Differentiate btw. callable and module not in env
Lately, we did not differentiate between a module not being in the
environment and functions not being in the environment and displayed the
same error message - "`this` is not part of the environment".

Adds logic to differentiate between callable not being present and
module not being present and display a different error if the module is
present but not the callable. This happens when a user writes a model
file with an invalid function/callable name.

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>