Remove `parent` field from assignment statements
Summary:
This was pretty straightforward. The only use was in
`typeCheck.ml`, but as was pointed out in the doc outlining
AST cleanups it was duplicated with the parent of the define,
which was already in scope. So I just use that parent instead.

It saves a good bit of boilerplate code and should also reduce
our RAM use a tiny bit.

Reviewed By: grievejia

Differential Revision: D31609565

fbshipit-source-id: 6fc305048651ef9b25b20028778a2d5c1e545c36