Add validation checks for taint.config files
Lately, we didn't validate taint.config files which led to some runs
yielding ambiguous or misleading results. Adds code to do this
verification before running the server on the analyze command.

taint_config_verifier consists of a verification function which takes in
a list of taint_models_path from the configuration, searches for
taint.config files, parses them and validates them.

Adds check to detect duplicate warning codes, invalid sinks and sources
in rules, and invalid regex expressions in implicit sources and sinks.
Adds a custom error with return code 9 that will be returned and a
custom exception that'll be raised if any errors are encountered.

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>