Rename the `key` parameter into `statement_key`
Summary:
When computing a fixpoint, we pass around a `key` parameter that is unique for a given statement in a control flow graph node. This is used to index the set of errors, and also to cache the annotations at a given point of the program.
This diff simply renames `key` into `statement_key`, as it seems like a better name and makes it easier to understand.
For instance, I have been confused for a while about this key parameter.

Reviewed By: kbansal

Differential Revision: D31693667

fbshipit-source-id: be206f101681f77268a23cda87b28e694f475d7b