Move errors and fixpoint resolution out of the type check fixpoint state
Summary:
The type check analysis uses to hashtables:
* The error map, which is a map from statement key to a list of errors
* The annotation map, which is a map from statement key to a set of annotations

These maps are currently part of the fixpoint state, but are not involved at all during join/widening and less_or_equal comparison.
Worst is that these are hashtables and thus have a reference semantic in OCaml. For instance, analyzing a node does update the error map and annotation map of other nodes. Actually, there is only one error map and annotation map allocated for the whole analysis.

This diff moves the error map and annotation map (called fixpoint resolution) from the fixpoint state to the fixpoint context, which is a more natural place for these. We also use an optional type here for analyses that do not need to store these along the way, which is the case for the taint analysis.
This is also necessary to define a proper `bottom` element, which will be required in the next diff.

Reviewed By: stroxler

Differential Revision: D31694852

fbshipit-source-id: d59a44505e48ba3032e2c350d2a01fc9d05ca836