Revive pyre infer global tests
Summary:
We lost the ability to run interprocedural infer when
`pyre analyze` moved to new-server. This is okay, we no
longer expect to ever use interprocedural infer (instead, we'll
probably rewrite the AST and use incremental plumbing if / when
we do cross-function inference; it's also not clear we'll even
get that far depending on some other projects).

But it was good to have tests of the logic for combining local
results. This diff uses `service/infer.ml` as an interface to
run the tests rather than interprocedural, and revives the test
data that was stripped out in D30196050 (https://github.com/facebook/pyre-check/commit/d7e301a4a017bb3f9e942275258d1cddc4921ac7)OB

Reviewed By: pradeep90

Differential Revision: D31731610

fbshipit-source-id: 4be60e908d5d4be921fd61121c10fa3dc647ca1b