Infer: add performance logging
Summary:
Pyre infer started becoming a significant bottleneck for pysa
about a month ago. Adding some per-define logging so we can
use scuba to hunt for the root cause.

It's not clear whehter we should sample logs, for now I decided
to log every 1 because I think pyre infer is run infrequently
enough - codemods are negligible, it's almost never run on
devservers, and we only run against a subset of fbcode diffs.

Reviewed By: arthaud

Differential Revision: D31734591

fbshipit-source-id: 385467414fc4497f0b85a75791c730a7c201a7bb