Infer: skip fully-annotated defines
Summary:
Skip analysis of a define in `pyre infer` unless one of the
following is true:
- the define is a constructor, class toplevel, or module toplevel
- the return annotation or some parameter annotation
  - is missing (no expression)
  - is untyped (literal Any, or the type fails to evaluate)

We may eventually want a flag to control this for codemods if we start
doing more aggressive cross-define inference. The need for filtering
comes mainly from pysa where we really can't afford to do as much
extra work.

For this reason and also because it's convenient to still be able to test
against fully-annotated functions, I tweaked the API so that we can
skip the skipping. But for now there's no need to hook that flag up
to the command-line interface.

Reviewed By: dkgi

Differential Revision: D31743862

fbshipit-source-id: 14b3081e267473a9663ef5bd5d136744abb8a583